+++ HISTORY.md
@@ -7,9 +7,10 @@
 - `Robot.getVK()` を `Robot.VK` へ変更
 - `System.exec` のエラー処理を強化
 - `System.run` のエラー処理を強化、引数用 `System.AppWinStype` 定数を追加、戻り値について`doc`コメントを記載
+- `System.BatchScript` に戻り値の仕様変更、また `UTF-8` で正しくバッチ実行できるように修正。 `UTF-16LE` には未対応であることを`doc`コメントに明記
 - `Dialog.popupMessage` の一部の引数の定数を `Dialog.POPUP_OPTION_TYPE` 内へ移動
 - `Dialog.popupMessage` の戻り値を `Dialog.POPUP_RETURN` 内へ移動
-- `StringComparator.DEFAULT` と `StringComparator.NATURAL` の入力引数を `string` から `any` 型へ改善
+- `Array.prototype.sort` で利用する `StringComparator.DEFAULT` と `StringComparator.NATURAL` について `any` 型でもソート出来るように改善
 - `console.log` など文字列出力メソッドにて配列型、オブジェクト型を渡した場合、内部データを表示するように変更
 - `ExtendsObject` と `ExtendsArray` 内の `toString` が既に定義されており不要なので削除し、`ExtendsObject.js` を削除